Public API Reference

System Catalog

PowerSystemCaseBuilder.SystemCatalog — Type

A container for a catalog of PowerSystems.System data sets

Example

SystemCatalog()

Returns the PowerSystemCaseBuilder.jl catalog

PowerSystemCaseBuilder.list_categories — Method

list_categories() -> Vector{DataType}

Returns a vector of SystemCategorys available in the PowerSystemCaseBuilder.jl SystemCatalog

PowerSystemCaseBuilder.list_systems — Method

list_systems(
    category::Type{<:SystemCategory}
) -> Vector{String}

Returns a vector of names of the systems available in one SystemCategory in the PowerSystemCaseBuilder.jl SystemCatalog

Reading the Catalog

PowerSystemCaseBuilder.show_categories — Method

show_categories()

Prints the names of the SystemCategorys available in the PowerSystemCaseBuilder.jl catalog

PowerSystemCaseBuilder.show_systems — Method

show_systems(category::Type{<:SystemCategory}; kwargs...)

Prints the name and description of all PowerSystems.Systems in a selected SystemCategory in the PowerSystemCaseBuilder.jl catalog

PowerSystemCaseBuilder.show_systems — Method

show_systems(; kwargs...)

Prints the name and description of all PowerSystems.Systems in the PowerSystemCaseBuilder.jl catalog, grouped by SystemCategory

See also: show_systems for a selected category

Categories of Systems

PowerSystemCaseBuilder.SystemCategory — Type

Supertype of categories that group test PowerSystems.Systems for similar use cases

Current subtypes with example Systems for running test simulations include:

PSISystems

PSIDSystems

Subtypes for testing other packages include:

PSYTestSystems

PSSEParsingTestSystems

MatpowerTestSystems

PSITestSystems

PSIDTestSystems

PowerSystemCaseBuilder.MatpowerTestSystems — Type

Category of PowerSystems.Systems to test parsing Matpower files.

Systems only include data for the power flow case.

PowerSystemCaseBuilder.PSIDSystems — Type

Category of example PowerSystems.Systems for running PowerSimulationsDynamics.jl simulations.

PowerSystemCaseBuilder.PSIDTestSystems — Type

Category of PowerSystems.Systems for PowerSimulationsDynamics.jl package testing.

Warning

Not all Systems are functional.

PowerSystemCaseBuilder.PSISystems — Type

Category of example PowerSystems.Systems for running PowerSimulations.jl operations problems and simulations.

PowerSystemCaseBuilder.PSITestSystems — Type

Category of PowerSystems.Systems for PowerSimulations.jl package testing.

Warning

Not all Systems are functional.

PowerSystemCaseBuilder.PSSEParsingTestSystems — Type

Category of PowerSystems.Systems to test parsing PSSe .raw files.

Systems only include data for the power flow case.

PowerSystemCaseBuilder.PSYTestSystems — Type

Category of PowerSystems.Systems for PowerSystems.jl package testing.

Warning

Not all Systems are functional.

PowerSystemCaseBuilder.SPISystems — Type

Category for SiennaPRASInterface.jl examples.

Building a System

PowerSystemCaseBuilder.build_system — Function

build_system(
    category::Type{<:SystemCategory},
    name::String;
    ...
) -> Any
build_system(
    category::Type{<:SystemCategory},
    name::String,
    print_stat::Bool;
    force_build,
    assign_new_uuids,
    skip_serialization,
    system_catalog,
    kwargs...
) -> Any

Builds a PowerSystems.System from one of the predefined categories of test systems

Arguments

category::Type{<:SystemCategory}: A subtype of SystemCategory

name::String: Name of the System to build

Accepted Key Words

print_stat::Bool = false: Print statistics about the system build process

force_build::Bool: true runs entire build process, false (Default) uses deserializiation if possible

assign_new_uuids::Bool: Assign new UUIDs to the system and all components if deserialization is used. Default is true.

skip_serialization::Bool: Default is false

system_catalog::SystemCatalog: Defaults to the PowerSystemCaseBuilder.jl catalog of Systems

Utilities

PowerSystemCaseBuilder.SystemDescriptor — Method

SystemDescriptor(
;
    name,
    description,
    category,
    build_function,
    raw_data,
    download_function,
    stats,
    supported_arguments
)

Constructs a container for describing, categorizing, and building a PowerSystems.System

Arguments

name::AbstractString: Unique name of the System

description::AbstractString: Phrase describing the System and key distinguishing features from other systems in the catalog

category::Type{<:SystemCategory}: A subtype of SystemCategory for categorizing this System within the catalog

raw_data::AbstractString: A path pointing to the System's data directory, .jl file, or parseable .m or .raw file in PowerSystemsTestData

build_function::Function: A custom Function for building this System. See src/library files.

download_function::Union{Nothing, Function}: Typically unused

stats::Union{Nothing,SystemBuildStats}: Container for summary statistics on the build process

supported_arguments::Vector{SystemArgument}: Additional arguments to pass to the build_function

PowerSystemCaseBuilder.SystemBuildStats — Method

SystemBuildStats(
;
    count,
    initial_construct_time,
    serialize_time,
    min_deserialize_time,
    max_deserialize_time,
    total_deserialize_time
)

Returns a container for the summary statistics on the build process for a PowerSystems.System