Internal API

PowerSystemCaseBuilder.SystemArgument — Type

An argument to pass to a SystemDescriptor's build_function

PowerSystemCaseBuilder.SystemArgument — Method

SystemArgument(; name, default, allowed_values)

Returns an argument to pass to a SystemDescriptor's build_function

Example

SystemArgument(;
                name = :add_forecasts,
                default = true,
                allowed_values = Set([true, false]),
            )

Downloads.download — Function

download(
    repo::AbstractString,
    branch::AbstractString,
    folder::AbstractString
) -> Any
download(
    repo::AbstractString,
    branch::AbstractString,
    folder::AbstractString,
    force::Bool
) -> Any

Download Data from a "branch" into a "data" folder in given argument path. Skip the actual download if the folder already exists and force=false. Returns the downloaded folder name.

PowerSystemCaseBuilder.build_pwl_average_cost_test_sys — Method

build_pwl_average_cost_test_sys(
;
    kwargs...
) -> PowerSystems.System

Same as buildpwlincrementalcosttest_sys but converts on the fly to PiecewiseAverageCurve

PowerSystemCaseBuilder.build_pwl_average_fuel_test_sys — Method

build_pwl_average_fuel_test_sys(
;
    kwargs...
) -> PowerSystems.System

Same as buildpwlincrementalfueltest_sys but converts on the fly to PiecewiseAverageCurve

PowerSystemCaseBuilder.clear_all_serialized_systems — Method

clear_all_serialized_systems(path::String)

Delete all data and files for PowerSystems.Systems that have been serialized

PowerSystemCaseBuilder.clear_serialized_system — Function

clear_serialized_system(name::String)
clear_serialized_system(
    name::String,
    case_args::Dict{Symbol}
)

Delete all files and references associated with a PowerSystems.System that has been serialized

Useful while debugging adding a new System to PowerSystemCaseBuilder.jl to ensure all data and references are removed and that the new build is clean

PowerSystemCaseBuilder.filter_descriptor_kwargs — Method

filter_descriptor_kwargs(
    descriptor::SystemDescriptor;
    kwargs...
) -> Dict{Symbol, Any}

Return the keyword arguments passed by the user that apply to the descriptor. Add any default values for fields not passed by the user.

PowerSystemCaseBuilder.get_total_system_count — Method

get_total_system_count(sys::SystemCatalog) -> Int64

Returns the total number of systems in a SystemCatalog

PowerSystemCaseBuilder.rate_to_probability — Method

rate_to_probability(
    for_gen::Float64,
    mttr::Int64
) -> @NamedTuple{λ::Float64, μ::Float64}

rate_to_probability(for_gen::Float64, mttr::Int64)

Converts the forced outage rate and mean time to repair to the λ and μ parameters

Arguments

for_gen::Float64: Forced outage rate [1/T]

mttr::Int64: Mean time to repair [T]

Returns

λ::Float64: Transition probability from online to offline [1/T]

μ::Float64: Transition rate from offline to online [1/T]

Reference

https://core.ac.uk/download/pdf/13643059.pdf from https://github.com/NREL-Sienna/SiennaPRASInterface.jl/blob/main/src/util/draws/drawhelperfunctions.jl

PowerSystemCaseBuilder.serialize_case_parameters — Method

serialize_case_parameters(case_args::Dict{Symbol}) -> Any

Creates a JSON file informing the user about the meaning of the hash value in the file path if it doesn't exist already